requirements.txt
uvicorn==0.12.1
fastapi==0.61.1
celery==4.4.7
redis==3.5.3
python-dotenv==0.14.0